Checking empty file before updating data

+++ update.sh
@@ -1,13 +1,23 @@
-#!/bin/bash  
+#!/bin/bash
 
-wget -O /users/altino/Sites/checkmarks/data/dados.json  "https://sistemas.anac.gov.br/dadosabertos/Aeronaves/RAB/dados_aeronaves.json"
+wget -O /users/altino/Sites/checkmarks/data/temp_dados.json  "https://sistemas.anac.gov.br/dadosabertos/Aeronaves/RAB/dados_aeronaves.json"
 
-var=`date +"%FORMAT_STRING"`
-now=`date +"%m_%d_%Y"`
-now=`date +"%d de %B de %Y"`
-echo "${now}" > update_date.txt
+if [ -s /users/altino/Sites/checkmarks/data/temp_dados.json ]
+then
+    var=`date +"%FORMAT_STRING"`
+    now=`date +"%m_%d_%Y"`
+    now=`date +"%d de %B de %Y"`
+    echo "${now}" > update_date.txt
+    
+    rm /users/altino/Sites/checkmarks/data/dados.json
+    mv /users/altino/Sites/checkmarks/data/temp_dados.json /users/altino/Sites/checkmarks/data/dados.json
 
-git add /users/altino/Sites/checkmarks/data/dados.json  
-git add /users/altino/Sites/checkmarks/data/update_date.txt  
-git commit -m "update data"
-git push origin main
+    git add /users/altino/Sites/checkmarks/data/dados.json
+    git add /users/altino/Sites/checkmarks/update_date.txt
+
+    git commit -m "update data"
+    git push origin main
+    
+else
+    echo "Problema ao baixar arquivo JSON do RAB. Nada foi alterado na base de dados."
+fi
